ref: lottoApplication 서비스 코드로 이동
